+++ CHANGELOG.md
@@ -20,9 +20,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   * nextSortedIntTriple: a variation of nextIntTriple whose result is in sorted order.
   * nextSortedWindowedIntTriple: a variation of nextWindowedIntTriple whose result is in sorted order.
 * Methods in the EnhancedRandomGenerator class for generating streams of combinations of distinct integers as streams of IndexPair or IndexTriple objects, including streams of the following:
-  * Random pairs of distinct integers
-  * Random triples of distinct integers
-  * Random triples of distinct sorted integers
+  * Random pairs of distinct integers (method pairs)
+  * Random pairs of distinct integers, separated by at most a specified window (method windowedPairs)
+  * Random triples of distinct integers (method triples)
+  * Random triples of distinct sorted integers (method sortedTriples)
 
 ### Changed
 * Refactored and optimized RandomIndexer.nextIntTriple methods.